requirements.txt
streamlit
gtts
SpeechRecognition 
pydub
openai-whisper
pytesseract
Pillow
pdf2image
langdetect